refactor(OutputView): move AttackType class
