+++ src/main/java/bossmonster/domain/BossMonster.java
@@ -1,7 +1,5 @@
 package bossmonster.domain;
 
-import bossmonster.AttackType;
-
 public class BossMonster {
 
     int hp;
